+++ src/plugins/storage-worker/workers/lokijs-indexeddb.worker.ts
@@ -1,5 +1,5 @@
-import { getRxStorageLoki } from '../../lokijs';
-import { wrappedWorkerRxStorage } from '../../worker';
+import { getRxStorageLoki } from '../../storage-lokijs';
+import { wrappedWorkerRxStorage } from '../../storage-worker';
 const LokiIndexedDBAdapter = require('lokijs/src/loki-indexed-adapter');
 
 const storage = getRxStorageLoki({