+++ src/plugins/worker/workers/dexie.worker.ts
@@ -1,7 +1,0 @@
-import { getRxStorageDexie } from '../../dexie';
-import { wrappedWorkerRxStorage } from '../../worker';
-
-const storage = getRxStorageDexie();
-wrappedWorkerRxStorage({
-    storage
-});