+++ test/unit/config.ts
@@ -8,14 +8,14 @@ import {
 import * as path from 'path';
 import parallel from 'mocha.parallel';
 import type { RxTestStorage } from '../../';
-import { getRxStorageLoki } from '../../plugins/lokijs';
+import { getRxStorageLoki } from '../../plugins/storage-lokijs';
 import {
     getRxStorageDexie,
     RxStorageDexieStatics
-} from '../../plugins/dexie';
-import { getRxStorageWorker } from '../../plugins/worker';
+} from '../../plugins/storage-dexie';
+import { getRxStorageWorker } from '../../plugins/storage-worker';
 import { getRxStorageRemoteWebsocket } from '../../plugins/storage-remote';
-import { getRxStorageMemory } from '../../plugins/memory';
+import { getRxStorageMemory } from '../../plugins/storage-memory';
 import { CUSTOM_STORAGE } from './custom-storage';
 import { wrappedValidateAjvStorage } from '../../plugins/validate-ajv';
 import { isPromise } from 'async-test-util';
@@ -215,7 +215,7 @@ export function setDefaultStorage(storageKey: string) {
             break;
         case 'dexie-worker':
             const dexieMemoryWorkerPath = require('path').join(
-                '../../../../dist/lib/plugins/worker/workers/',
+                '../../../../dist/lib/plugins/storage-worker/workers/',
                 'dexie-memory.worker.js'
             );
             console.log('dexieMemoryWorkerPath: ' + dexieMemoryWorkerPath);
@@ -249,7 +249,7 @@ export function setDefaultStorage(storageKey: string) {
             const foundationDBAPIVersion = 620;
 
             // use a dynamic import so it does not break browser bundling
-            const { getRxStorageFoundationDB } = require('../../plugins/foundationdb' + '');
+            const { getRxStorageFoundationDB } = require('../../plugins/storage-foundationdb' + '');
 
             config.storage = {
                 name: storageKey,