+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 - CHANGE `RxCollection.findByIds()` now returns a `RxQuery`.
 - REMOVED `RxCollection.findByIds$`, use `RxCollection.findByIds().$` instead.
 - FIX schema hashing should respect the sort order [#4005](https://github.com/pubkey/rxdb/pull/4005)
+- CHANGE Prefix storage plugins with `storage-` like `rxdb/plugins/storage-dexie`.
+
 
 <!-- ADD new changes here! -->